+++ apscheduler/schedulers/qt.py
@@ -9,13 +9,16 @@ except (ImportError, RuntimeError):  # pragma: nocover
         from PyQt4.QtCore import QObject, QTimer
     except ImportError:
         try:
-            from PySide2.QtCore import QObject, QTimer  # noqa
+            from PySide6.QtCore import QObject, QTimer  # noqa
         except ImportError:
             try:
-                from PySide.QtCore import QObject, QTimer  # noqa
+                from PySide2.QtCore import QObject, QTimer  # noqa
             except ImportError:
-                raise ImportError('QtScheduler requires either PyQt5, PyQt4, PySide2 '
-                                  'or PySide installed')
+                try:
+                    from PySide.QtCore import QObject, QTimer  # noqa
+                except ImportError:
+                    raise ImportError('QtScheduler requires either PyQt5, PyQt4, PySide6, PySide2 '
+                                      'or PySide installed')
 
 
 class QtScheduler(BaseScheduler):